+++ bin/check_compat.js
@@ -17,7 +17,8 @@ E.is_env_compat = ()=>{
     const pkg = require('../package.json');
     const excluded = ['angular', 'angular-sanitize', 'bootstrap',
         'bootstrap-datepicker', 'codemirror', 'notosans-fontface',
-        'require-css', 'flag-icon-css', 'angular-ui-bootstrap', 'swagger-ui']
+        'require-css', 'flag-icon-css', 'angular-ui-bootstrap',
+        'swagger-ui', 'font-awesome']
         .concat(Object.keys(pkg.optionalDependencies));
     for (let dep in pkg.dependencies)
     {