+++ db/update.js
@@ -2,20 +2,18 @@
 // LICENSE_CODE ZON
 'use strict'; /*jslint node:true*/
 
+const fs = require('fs');
 const yauzl = require('yauzl');
 const request = require('request');
-const hutil = require('hutil');
 const StringDecoder = require('string_decoder').StringDecoder;
-const sqlite3 = require('sqlite3');
-const file = hutil.file;
 
 const DB_URL = 'http://geolite.maxmind.com/download/geoip/database/GeoLite2-City-CSV.zip';
 const FILE_REGEX = /\/GeoLite2-City-Locations-en.csv$/;
 
 let unzip_file = (buf, mask, cb)=>{
-    yauzl.fromBuffer(buf, {lazyEntries: true}, (err, zip)=>{
-        if (err)
-            throw err;
+    yauzl.fromBuffer(buf, {lazyEntries: true}, (error, zip)=>{
+        if (error)
+            throw error;
         zip.readEntry();
         zip.on('entry', entry=>{
             if (!mask.test(entry.fileName))
@@ -57,66 +55,50 @@ let process_city_csv = s=>{
         return arr;
     }, []);
     console.log('+ csv parsed');
-    if (file.exists('./city.db'))
-        file.unlink_e('./city.db');
-    let db = new sqlite3.Database('./city.db');
-    db.serialize(()=>{
-        let fallback = function(err){
-            if (err)
-                throw err;
-        };
-        db.run('BEGIN TRANSACTION', [], fallback);
-        db.run('CREATE TABLE cities (name VARCHAR, region_id CHAR(3), '+
-            'country_id CHAR(2))', [], fallback);
-        db.run('CREATE TABLE countries (country_id CHAR(2), name VARCHAR, '+
-            'PRIMARY KEY (country_id))', [], fallback);
-        db.run('CREATE TABLE regions (region_id CHAR(3), name VARCHAR, '+
-            'country_id CHAR(2))', [], fallback);
-        db.run('CREATE INDEX region_country_id ON regions(country_id)', [],
-            fallback);
-        db.run('CREATE INDEX city_region_id ON cities(region_id)', [],
-            fallback);
-        db.run('CREATE INDEX city_country_id ON cities(country_id)', [],
-            fallback);
-        console.log('+ schema created');
-        let stmt, c = 0;
-        db.parallelize(()=>{
-            stmt = db.prepare('INSERT INTO countries VALUES (?, ?)', [],
-                fallback);
-            for (let country_id in countries)
-            {
-                stmt.run(country_id, countries[country_id], fallback);
-                c++;
-            }
-            console.log('+ '+c+' records in table "countries" added');
-        });
-        stmt.finalize();
-        db.parallelize(()=>{
-            c = 0;
-            stmt = db.prepare('INSERT INTO regions VALUES (?, ?, ?)', [],
-                fallback);
-            for (let country_id in regions)
-            {
-                for (let region_id in regions[country_id])
-                {
-                    stmt.run(region_id, regions[country_id][region_id],
-                        country_id, fallback);
-                    c++;
-                }
-            }
-            console.log('+ '+c+' records in table "regions" added');
-        });
-        stmt.finalize();
-        db.parallelize(()=>{
-            stmt = db.prepare('INSERT INTO cities VALUES (?, ?, ?)', [],
-                fallback);
-            csv_data.forEach(rec=>stmt.run(rec[10], rec[6], rec[4], fallback));
-            console.log('+ '+csv_data.length+' records in table cities added');
-        });
-        stmt.finalize();
-        db.run('END TRANSACTION', [], fallback);
+    let data = [], for_db = str=>str.replace(/'/g, "''");
+    data.push('BEGIN TRANSACTION;');
+    data.push('CREATE TABLE cities (name VARCHAR, region_id CHAR(3), '+
+        'country_id CHAR(2));');
+    csv_data.forEach(rec=>{
+        let name = for_db(rec[10]);
+        let r_id = for_db(rec[6]);
+        let c_id = for_db(rec[4]);
+        data.push(
+            `INSERT INTO "cities" VALUES('${name}','${r_id}','${c_id}');`);
     });
-    db.close();
+    console.log('+ '+csv_data.length+' records of cities added');
+    data.push('CREATE TABLE countries (country_id CHAR(2), name VARCHAR, '+
+        'PRIMARY KEY (country_id));');
+    for (let country_id in countries)
+    {
+        let id = for_db(country_id);
+        let name = for_db(countries[country_id]);
+        data.push(`INSERT INTO "countries" VALUES('${id}','${name}');`);
+    }
+    console.log(
+        '+ '+Object.keys(countries).length+' records of countries added');
+    let c = 0;
+    data.push('CREATE TABLE regions (region_id CHAR(3), name VARCHAR, '+
+        'country_id CHAR(2));');
+    for (let country_id in regions)
+    {
+        for (let region_id in regions[country_id])
+        {
+            let id = for_db(region_id);
+            let name = for_db(regions[country_id][region_id]);
+            let c_id = for_db(country_id);
+            data.push(
+                `INSERT INTO "regions" VALUES('${id}','${name}','${c_id}');`);
+            c++;
+        }
+    }
+    console.log('+ '+c+' records of regions added');
+    data.push('CREATE INDEX region_country_id ON regions(country_id);');
+    data.push('CREATE INDEX city_region_id ON cities(region_id);');
+    data.push('CREATE INDEX city_country_id ON cities(country_id);');
+    data.push('COMMIT;');
+    data.push('');
+    fs.writeFileSync('./city.sql', data.join('\n'));
 };
 
 request(DB_URL, {encoding: null}, (err, res, body)=>{